separate layer in modules (#72)
* separate layer in modules

* wip

* compiling

* tests passing

* full refactor
