+++ zkml/src/layers/activation.rs
@@ -1,0 +1,277 @@
+use crate::{
+    Claim, Prover,
+    commit::same_poly,
+    iop::{context::ContextAux, verifier::Verifier},
+    layers::{LayerCtx, LayerProof, PolyID},
+    lookup::{
+        context::TableType,
+        logup_gkr::{
+            prover::batch_prove as logup_batch_prove, structs::LogUpProof,
+            verifier::verify_logup_proof,
+        },
+    },
+};
+use ff_ext::ExtensionField;
+use gkr::util::ceil_log2;
+use multilinear_extensions::mle::IntoMLE;
+use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
+use serde::{Deserialize, Serialize, de::DeserializeOwned};
+use transcript::Transcript;
+
+use crate::{
+    Element,
+    quantization::{self, BIT_LEN, Fieldizer},
+    tensor::Tensor,
+};
+
+#[derive(Clone, Debug, Serialize, Deserialize, Copy)]
+pub enum Activation {
+    Relu(Relu),
+}
+
+/// Currently holds the poly info for the output polynomial of the RELU
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ActivationCtx {
+    pub op: Activation,
+    pub poly_id: PolyID,
+    pub num_vars: usize,
+}
+
+#[derive(Clone, Serialize, Deserialize)]
+pub struct ActivationProof<E: ExtensionField>
+where
+    E::BaseField: Serialize + DeserializeOwned,
+{
+    /// proof for the accumulation of the claim from m2v + claim from lookup for the same poly
+    /// e.g. the "link" between a m2v and relu layer
+    pub(crate) io_accumulation: same_poly::Proof<E>,
+    /// the lookup proof for the relu
+    pub(crate) lookup: LogUpProof<E>,
+}
+
+impl Activation {
+    pub fn op(&self, input: &Tensor<Element>) -> Tensor<Element> {
+        match self {
+            Activation::Relu(relu) => relu.op(input),
+        }
+    }
+    pub(crate) fn step_info<E: ExtensionField>(
+        &self,
+        id: PolyID,
+        mut aux: ContextAux,
+    ) -> (LayerCtx<E>, ContextAux)
+    where
+        E: ExtensionField + DeserializeOwned,
+        E::BaseField: Serialize + DeserializeOwned,
+    {
+        aux.tables.insert(TableType::Relu);
+        let info = match self {
+            Activation::Relu(relu) => LayerCtx::Activation(ActivationCtx {
+                op: Activation::Relu(*relu),
+                poly_id: id,
+                num_vars: aux
+                    .last_output_shape
+                    .iter()
+                    .map(|dim| ceil_log2(*dim))
+                    .sum::<usize>(),
+            }),
+        };
+        (info, aux)
+    }
+
+    pub(crate) fn prove_step<E: ExtensionField, T: Transcript<E>>(
+        &self,
+        prover: &mut Prover<E, T>,
+        last_claim: &Claim<E>,
+        output: &[E],
+        step: &ActivationCtx,
+    ) -> anyhow::Result<Claim<E>>
+    where
+        E: ExtensionField + Serialize + DeserializeOwned,
+        E::BaseField: Serialize + DeserializeOwned,
+    {
+        let prover_info = prover.next_lookup_witness()?;
+
+        // Run the lookup protocol and return the lookup proof
+        let logup_proof = logup_batch_prove(&prover_info, prover.transcript)?;
+
+        // We need to prove that the output of this step is the input to following activation function
+        let mut same_poly_prover = same_poly::Prover::<E>::new(output.to_vec().into_mle());
+        let same_poly_ctx = same_poly::Context::<E>::new(last_claim.point.len());
+        same_poly_prover.add_claim(last_claim.clone())?;
+        // Activation proofs have two columns, input and output
+        let input_claim = logup_proof.output_claims()[0].clone();
+        let output_claim = logup_proof.output_claims()[1].clone();
+
+        same_poly_prover.add_claim(output_claim)?;
+        let claim_acc_proof = same_poly_prover.prove(&same_poly_ctx, prover.transcript)?;
+        // order is (output,mult)
+        prover
+            .witness_prover
+            .add_claim(step.poly_id, claim_acc_proof.extract_claim())?;
+
+        // Add the proof in
+        prover.push_proof(LayerProof::Activation(ActivationProof {
+            io_accumulation: claim_acc_proof,
+            lookup: logup_proof,
+        }));
+        Ok(input_claim)
+    }
+}
+
+impl ActivationCtx {
+    pub(crate) fn verify_activation<E: ExtensionField, T: Transcript<E>>(
+        &self,
+        verifier: &mut Verifier<E, T>,
+        last_claim: Claim<E>,
+        proof: &ActivationProof<E>,
+        constant_challenge: E,
+        column_separation_challenge: E,
+    ) -> anyhow::Result<Claim<E>>
+    where
+        E::BaseField: Serialize + DeserializeOwned,
+        E: Serialize + DeserializeOwned,
+    {
+        // 1. Verify the lookup proof
+        let verifier_claims = verify_logup_proof(
+            &proof.lookup,
+            1,
+            constant_challenge,
+            column_separation_challenge,
+            verifier.transcript,
+        )?;
+
+        // 2. Verify the accumulation proof from last_claim + lookup claim into the new claim
+        let sp_ctx = same_poly::Context::<E>::new(self.num_vars);
+        let mut sp_verifier = same_poly::Verifier::<E>::new(&sp_ctx);
+        sp_verifier.add_claim(last_claim)?;
+        verifier_claims.claims()[1..]
+            .iter()
+            .try_for_each(|claim| sp_verifier.add_claim(claim.clone()))?;
+
+        let new_output_claim = sp_verifier.verify(&proof.io_accumulation, verifier.transcript)?;
+        // 3. Accumulate the new claim into the witness commitment protocol
+        verifier
+            .witness_verifier
+            .add_claim(self.poly_id, new_output_claim)?;
+
+        // 4. return the input claim for to be proven at subsequent step
+        Ok(verifier_claims.claims()[0].clone())
+    }
+}
+
+#[derive(Clone, Debug, Copy, Serialize, Deserialize)]
+pub struct Relu;
+
+impl Relu {
+    pub fn new() -> Relu {
+        Self
+    }
+    pub fn num_vars() -> usize {
+        *BIT_LEN
+    }
+    pub fn poly_len() -> usize {
+        1 << Self::num_vars()
+    }
+    pub fn shape() -> Vec<usize> {
+        vec![2, Self::poly_len()]
+    }
+    /// to_mle returns two polynomials:
+    /// f_i: one containing the input column values
+    /// f_o: one containing the output column values
+    pub fn to_mle<E: ExtensionField>() -> (Vec<E::BaseField>, Vec<E::BaseField>) {
+        (*quantization::MIN..=*quantization::MAX)
+            .map(|i| {
+                let val: E = i.to_field();
+                let op_val: E = Relu::apply(i as i128).to_field();
+                (val.as_bases()[0], op_val.as_bases()[0])
+            })
+            .unzip()
+    }
+
+    pub fn op(&self, input: &Tensor<Element>) -> Tensor<Element> {
+        Tensor::new(
+            input.dims(),
+            input
+                .get_data()
+                .par_iter()
+                .map(|e| Self::apply(*e))
+                .collect::<Vec<_>>(),
+        )
+    }
+
+    #[inline(always)]
+    pub fn apply(e: Element) -> Element {
+        if e.is_negative() { 0 } else { e }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use crate::to_bit_sequence_le;
+    use goldilocks::GoldilocksExt2;
+    use itertools::Itertools;
+    use multilinear_extensions::mle::{DenseMultilinearExtension, MultilinearExtension};
+
+    use super::*;
+
+    type F = GoldilocksExt2;
+
+    #[test]
+    fn test_activation_relu_apply() {
+        struct TestCase {
+            input: Element,
+            output: Element,
+        }
+
+        impl TestCase {
+            pub fn from(input: Element, output: Element) -> Self {
+                Self { input, output }
+            }
+        }
+        for case in [
+            TestCase::from(-24, 0),
+            TestCase::from(0, 0),
+            TestCase::from(124, 124),
+        ] {
+            assert_eq!(Relu::apply(case.input), case.output);
+        }
+    }
+
+    #[test]
+    fn test_activation_relu_mle() {
+        let relu = Relu::new();
+        let (input_poly, output_poly) = Relu::to_mle::<F>();
+
+        assert_eq!(input_poly.len(), output_poly.len());
+        let (input_mle, output_mle) = (
+            DenseMultilinearExtension::from_evaluation_vec_smart(
+                Relu::num_vars(),
+                input_poly.to_vec(),
+            ),
+            DenseMultilinearExtension::from_evaluation_vec_smart(
+                Relu::num_vars(),
+                output_poly.to_vec(),
+            ),
+        );
+        assert_eq!(input_mle.num_vars(), output_mle.num_vars());
+        assert_eq!(input_mle.num_vars(), Relu::num_vars());
+        let inputs = Tensor::random(vec![10]);
+        let outputs = relu.op(&inputs);
+        assert_eq!(inputs.dims(), outputs.dims());
+        for (input, output) in inputs.get_data().iter().zip(outputs.get_data().iter()) {
+            // here putting input works because every random input is a u8, so it's already within [0;256] so
+            // its value "is" the index. Normally if this is not true, we should get the index of the row corresponding to that input
+            let idx_vars = to_bit_sequence_le((input + 128) as usize, Relu::num_vars())
+                .map(|b| F::from(b as u64))
+                .collect_vec();
+            let input_field = input_mle.evaluate(&idx_vars);
+            let expected_ified: F = input.to_field();
+            assert_eq!(input_field, expected_ified);
+            let output_field = output_mle.evaluate(&idx_vars);
+            let expected_ofield: F = output.to_field();
+            assert_eq!(output_field, expected_ofield);
+        }
+        // assert_eq!(expected,given);
+    }
+}
