+++ zkml/src/layers/pooling.rs
@@ -1,20 +1,68 @@
 use crate::{
-    Element,
+    Claim, Element, Prover,
+    commit::{compute_betas_eval, identity_eval, precommit::PolyID, same_poly},
+    iop::verifier::Verifier,
+    layers::{ContextAux, LayerProof},
+    lookup::{
+        context::TableType,
+        logup_gkr::{
+            prover::batch_prove as logup_batch_prove, structs::LogUpProof,
+            verifier::verify_logup_proof,
+        },
+    },
     quantization::{Fieldizer, IntoElement},
     tensor::Tensor,
 };
+use anyhow::{Context, ensure};
 use ff_ext::ExtensionField;
 use gkr::util::ceil_log2;
 use itertools::{Itertools, izip};
+use multilinear_extensions::{
+    mle::{ArcDenseMultilinearExtension, DenseMultilinearExtension, IntoMLE, MultilinearExtension},
+    virtual_poly::{ArcMultilinearExtension, VPAuxInfo, VirtualPolynomial},
+};
+use serde::de::DeserializeOwned;
+use sumcheck::structs::{IOPProof, IOPVerifierState};
+use transcript::Transcript;
 
 use rayon::prelude::*;
 use serde::{Deserialize, Serialize};
+use sumcheck::structs::IOPProverState;
+
+use super::LayerCtx;
 
 pub const MAXPOOL2D_KERNEL_SIZE: usize = 2;
 
 #[derive(Clone, Debug, Serialize, Deserialize, Copy)]
 pub enum Pooling {
     Maxpool2D(Maxpool2D),
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct PoolingCtx {
+    pub poolinfo: Maxpool2D,
+    pub poly_id: PolyID,
+    pub num_vars: usize,
+}
+
+/// Contains proof material related to one step of the inference
+#[derive(Clone, Serialize, Deserialize)]
+pub struct PoolingProof<E: ExtensionField>
+where
+    E::BaseField: Serialize + DeserializeOwned,
+{
+    /// the actual sumcheck proof proving that the product of correct terms is always zero
+    pub(crate) sumcheck: IOPProof<E>,
+    /// The lookup proof showing that the diff is always in the correct range
+    pub(crate) lookup: LogUpProof<E>,
+    /// proof for the accumulation of the claim from the zerocheck + claim from lookup for the same poly for both input and output
+    pub(crate) io_accumulation: same_poly::Proof<E>,
+    /// The claims that are accumulated for the output of this step
+    pub(crate) output_claims: Vec<Claim<E>>,
+    /// The output evaluations of the diff polys produced by the zerocheck
+    pub(crate) zerocheck_evals: Vec<E>,
+    /// This tells the verifier how far apart the variables get fixed on the input MLE
+    pub(crate) variable_gap: usize,
 }
 
 impl Pooling {
@@ -24,6 +72,38 @@ impl Pooling {
         }
     }
 
+    pub(crate) fn step_info<E: ExtensionField>(
+        &self,
+        id: PolyID,
+        mut aux: ContextAux,
+    ) -> (LayerCtx<E>, ContextAux)
+    where
+        E: ExtensionField + DeserializeOwned,
+        E::BaseField: Serialize + DeserializeOwned,
+    {
+        let info = match self {
+            Pooling::Maxpool2D(info) => {
+                aux.tables.insert(TableType::Range);
+                // Pooling only affects the last two dimensions
+                let total_number_dims = aux.last_output_shape.len();
+
+                aux.last_output_shape
+                    .iter_mut()
+                    .skip(total_number_dims - 2)
+                    .for_each(|dim| *dim = (*dim - info.kernel_size) / info.stride + 1);
+                LayerCtx::Pooling(PoolingCtx {
+                    poolinfo: *info,
+                    poly_id: id,
+                    num_vars: aux
+                        .last_output_shape
+                        .iter()
+                        .map(|dim| ceil_log2(*dim))
+                        .sum::<usize>(),
+                })
+            }
+        };
+        (info, aux)
+    }
     pub fn gen_lookup_witness<E: ExtensionField>(
         &self,
         input: &Tensor<Element>,
@@ -45,6 +125,283 @@ impl Pooling {
                 (merged_lookups, field_vecs)
             }
         }
+    }
+    #[timed::timed_instrument(level = "debug")]
+    pub fn prove_pooling<E: ExtensionField, T: Transcript<E>>(
+        &self,
+        prover: &mut Prover<E, T>,
+        // last random claim made
+        last_claim: Claim<E>,
+        // input to the dense layer
+        input: &Tensor<E>,
+        // output of dense layer evaluation
+        output: &Tensor<E>,
+        info: &PoolingCtx,
+    ) -> anyhow::Result<Claim<E>>
+    where
+        E::BaseField: Serialize + DeserializeOwned,
+        E: Serialize + DeserializeOwned,
+    {
+        assert_eq!(input.dims().len(), 3, "Maxpool needs 3D inputs.");
+        // Create the range check proof for the diff
+        let prover_info = prover.next_lookup_witness()?;
+
+        let logup_proof = logup_batch_prove(&prover_info, prover.transcript)?;
+
+        // These are the polys that get passed to the zero check make sure their product is zero at every evaluation point
+        let mut diff_polys = prover_info
+            .column_evals()
+            .iter()
+            .map(|diff| {
+                DenseMultilinearExtension::<E>::from_evaluations_slice(info.num_vars, diff).into()
+            })
+            .collect::<Vec<ArcMultilinearExtension<E>>>();
+
+        // Run the Zerocheck that checks enforces that output does contain the maximum value for the kernel
+        let mut vp = VirtualPolynomial::<E>::new(info.num_vars);
+
+        // We reuse the logup point here for the zerocheck challenge
+        let lookup_point = &logup_proof.output_claims()[0].point;
+
+        // Comput the identity poly
+        let batch_challenge = prover
+            .transcript
+            .get_and_append_challenge(b"batch_pooling")
+            .elements;
+
+        let beta_eval = compute_betas_eval(&lookup_point);
+        let beta_poly: ArcDenseMultilinearExtension<E> =
+            DenseMultilinearExtension::<E>::from_evaluations_ext_vec(info.num_vars, beta_eval)
+                .into();
+        let mut challenge_combiner = batch_challenge;
+        let lookup_parts = diff_polys
+            .iter()
+            .map(|diff| {
+                let out = (vec![diff.clone(), beta_poly.clone()], challenge_combiner);
+                challenge_combiner *= batch_challenge;
+                out
+            })
+            .collect::<Vec<(Vec<_>, E)>>();
+
+        diff_polys.push(beta_poly);
+        vp.add_mle_list(diff_polys, E::ONE);
+        lookup_parts
+            .into_iter()
+            .for_each(|(prod, coeff)| vp.add_mle_list(prod, coeff));
+
+        #[allow(deprecated)]
+        let (proof, sumcheck_state) = IOPProverState::<E>::prove_parallel(vp, prover.transcript);
+
+        // We need to prove that the output of this step is the input to following activation function
+        let output_mle = output.get_data().to_vec().into_mle();
+        let mut same_poly_prover = same_poly::Prover::<E>::new(output_mle.clone());
+
+        let zerocheck_point = &proof.point;
+        let output_zerocheck_eval = output_mle.evaluate(zerocheck_point);
+
+        // Accumulate claims about the output polynomial in each of the protocols we ran together with the final claim from the previous proof.
+        let mut output_claims = Vec::<Claim<E>>::new();
+        let same_poly_ctx = same_poly::Context::<E>::new(last_claim.point.len());
+        same_poly_prover.add_claim(last_claim.clone())?;
+
+        let zerocheck_claim = Claim {
+            point: zerocheck_point.clone(),
+            eval: output_zerocheck_eval,
+        };
+        same_poly_prover.add_claim(zerocheck_claim.clone())?;
+        output_claims.push(zerocheck_claim);
+
+        // This is the proof for the output poly
+        let claim_acc_proof = same_poly_prover.prove(&same_poly_ctx, prover.transcript)?;
+
+        let output_claim = claim_acc_proof.extract_claim();
+
+        prover
+            .witness_prover
+            .add_claim(info.poly_id, output_claim)
+            .context("unable to add claim")?;
+        // Now we must do the samething accumulating evals for the input poly as we fix variables on the input poly.
+        // The point length is 2 longer because for now we only support MaxPool2D.
+
+        let padded_input_shape = input.dims();
+        let padded_input_row_length_log = ceil_log2(padded_input_shape[2]);
+        // We can batch all of the claims for the input poly with 00, 10, 01, 11 fixed into one with random challenges
+        let [r1, r2] = [prover
+            .transcript
+            .get_and_append_challenge(b"input_batching")
+            .elements; 2];
+
+        let one_minus_r1 = E::ONE - r1;
+        let one_minus_r2 = E::ONE - r2;
+        // To the input claims we add evaluations at both the zerocheck point and lookup point
+        // in the order 00, 01, 10, 11. These will be used in conjunction with r1 and r2 by the verifier to link the claims output by the sumcheck and lookup GKR
+        // proofs with the claims fed to the same poly verifier.
+
+        let multiplicands = [
+            one_minus_r1 * one_minus_r2,
+            one_minus_r1 * r2,
+            r1 * one_minus_r2,
+            r1 * r2,
+        ];
+
+        let zc_in_claim = izip!(
+            multiplicands.iter(),
+            sumcheck_state.get_mle_final_evaluations().iter(),
+        )
+        .fold(E::ZERO, |zc_acc, (m, zc)| {
+            zc_acc + *m * (output_zerocheck_eval - *zc)
+        });
+
+        let point_1 = [
+            &[r1],
+            &zerocheck_point[..padded_input_row_length_log - 1],
+            &[r2],
+            &zerocheck_point[padded_input_row_length_log - 1..],
+        ]
+        .concat();
+
+        let next_claim = Claim {
+            point: point_1,
+            eval: zc_in_claim,
+        };
+
+        // We don't need the last eval of the the sumcheck state as it is the beta poly
+
+        let zerocheck_evals = sumcheck_state.get_mle_final_evaluations()[..4].to_vec();
+        // Push the step proof to the list
+        prover.push_proof(LayerProof::Pooling(PoolingProof {
+            sumcheck: proof,
+            lookup: logup_proof,
+            io_accumulation: claim_acc_proof,
+            output_claims,
+            zerocheck_evals,
+            variable_gap: padded_input_row_length_log - 1,
+        }));
+        Ok(next_claim)
+    }
+}
+
+impl PoolingCtx {
+    pub(crate) fn verify_pooling<E: ExtensionField, T: Transcript<E>>(
+        &self,
+        verifier: &mut Verifier<E, T>,
+        last_claim: Claim<E>,
+        proof: &PoolingProof<E>,
+        constant_challenge: E,
+        column_separation_challenge: E,
+    ) -> anyhow::Result<Claim<E>>
+    where
+        E::BaseField: Serialize + DeserializeOwned,
+        E: Serialize + DeserializeOwned,
+    {
+        // 1. Verify the lookup proof
+        let verifier_claims = verify_logup_proof(
+            &proof.lookup,
+            4,
+            constant_challenge,
+            column_separation_challenge,
+            verifier.transcript,
+        )?;
+
+        // 2. Verify the sumcheck proof
+        let poly_aux = VPAuxInfo::<E>::from_mle_list_dimensions(&[vec![self.num_vars; 5]]);
+        let batching_challenge = verifier
+            .transcript
+            .get_and_append_challenge(b"batch_pooling")
+            .elements;
+        let initial_value = verifier_claims
+            .claims()
+            .iter()
+            .fold((E::ZERO, batching_challenge), |(acc, comb), claim| {
+                (acc + claim.eval * comb, comb * batching_challenge)
+            })
+            .0;
+        let subclaim = IOPVerifierState::<E>::verify(
+            initial_value,
+            &proof.sumcheck,
+            &poly_aux,
+            verifier.transcript,
+        );
+
+        // Run the same poly verifier for the output claims
+        let sp_ctx = same_poly::Context::<E>::new(self.num_vars);
+        let mut sp_verifier = same_poly::Verifier::<E>::new(&sp_ctx);
+
+        sp_verifier.add_claim(last_claim)?;
+
+        let output_claims = &proof.output_claims;
+        output_claims
+            .iter()
+            .try_for_each(|claim| sp_verifier.add_claim(claim.clone()))?;
+
+        let output_proof = &proof.io_accumulation;
+
+        let commit_claim = sp_verifier.verify(output_proof, verifier.transcript)?;
+
+        // Add the result of the same poly verifier to the commitment verifier.
+        verifier
+            .witness_verifier
+            .add_claim(self.poly_id, commit_claim)?;
+
+        // Challenegs used to batch input poly claims together and link them with zerocheck and lookup verification output
+        let [r1, r2] = [verifier
+            .transcript
+            .get_and_append_challenge(b"input_batching")
+            .elements; 2];
+        let one_minus_r1 = E::ONE - r1;
+        let one_minus_r2 = E::ONE - r2;
+
+        let eval_multiplicands = [
+            one_minus_r1 * one_minus_r2,
+            one_minus_r1 * r2,
+            r1 * one_minus_r2,
+            r1 * r2,
+        ];
+        let zc_point = subclaim
+            .point
+            .iter()
+            .map(|chal| chal.elements)
+            .collect::<Vec<E>>();
+        let zerocheck_point = [
+            &[r1],
+            &zc_point[..proof.variable_gap],
+            &[r2],
+            &zc_point[proof.variable_gap..],
+        ]
+        .concat();
+
+        let zerocheck_input_eval = izip!(proof.zerocheck_evals.iter(), eval_multiplicands.iter())
+            .fold(E::ZERO, |zerocheck_acc, (&ze, &me)| {
+                zerocheck_acc + (output_claims[0].eval - ze) * me
+            });
+
+        let out_claim = Claim {
+            point: zerocheck_point,
+            eval: zerocheck_input_eval,
+        };
+
+        // Now we check consistency between the lookup/sumcheck proof claims and the claims passed to the same poly verifiers.
+        let beta_eval = identity_eval(&output_claims[0].point, &verifier_claims.claims()[0].point);
+
+        let computed_zerocheck_claim = proof
+            .zerocheck_evals
+            .iter()
+            .chain(std::iter::once(&beta_eval))
+            .product::<E>()
+            + proof
+                .zerocheck_evals
+                .iter()
+                .fold((E::ZERO, batching_challenge), |(acc, comb), v| {
+                    (acc + *v * beta_eval * comb, comb * batching_challenge)
+                })
+                .0;
+
+        ensure!(
+            computed_zerocheck_claim == subclaim.expected_evaluation,
+            "Computed zerocheck claim did not line up with output of sumcheck verification"
+        );
+
+        Ok(out_claim)
     }
 }
